#001 ajout fonctionnalité de réservation + bouton de réservation

+++ liborrow-webservice/liborrow-webservice-consumer/src/main/java/org/liborrow/webservice/consumer/contract/dao/WaitingListDao.java
@@ -2,6 +2,6 @@ package org.liborrow.webservice.consumer.contract.dao;
 
 public interface WaitingListDao {
 	
-	Integer getWaitingListSize(Long itemId);
+	Long getWaitingListSize(Long itemId);
 	Integer getLastPosition(Long itemId);
 }
